[android] Disable the screen wake lock by default
Organic Maps drains the battery like crazy with default settings.

**Steps to reproduce:**

1. Install and run the app.
2. Lock your device and put it on a flat surface, with the screen facing up.
3. Pick up the device OR just tap on the screen.
4. The screen will automatically turn on, this is an Android feature.
5. Organic Maps will never allow the screen to go off after that.

Now, imaging this case happens while the phone is in a pocket. The battery may
The battery will completely drain before the user even notice that screen is on.
I see Organic Maps draining battery and burning screens on my devices regulary.

The problem is caused by two options which are enabled by default -
"Show on the lock screen" and "Allow the screen to sleep". The first option
adds the app to the lock screen (which is weird on its own). The second
option doesn't allow the phone's screen to go off.

This patch does the following:

- Inverts "Allow the screen to sleep" option to "Keep the screen on".
- Sets the default value of the option to OFF (=allow screen to sleep).
- Applies the option only in FOLLOW & FOLLOW_ROTATE location modes.
- Forces the screen to keep on during navigation no matter of the option.

To summarize, with enabled "Keep the screen on" the screen will be kept
on in the following cases:

- During navigation.
- FOLLOW and FOLLOW_ROTATE location modes.

Without the option, the screen will be kept on only during navigation.

Signed-off-by: Roman Tsisyk <roman@tsisyk.com>